refactor(entry-points): drop legacy ml_playground.prepare shim
