+++ ml_playground/experiments/protocol.py
@@ -4,8 +4,7 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import Protocol
 
-from ml_playground.prepare import PreparerConfig
-from ml_playground.configuration import TrainerConfig, SamplerConfig
+from ml_playground.configuration import PreparerConfig, TrainerConfig, SamplerConfig
 
 
 @dataclass(frozen=True)
