+++ ml_playground/experiments/shakespeare/preparer.py
@@ -4,13 +4,13 @@ from pathlib import Path
 import numpy as np
 import requests
 import requests.exceptions
-from ml_playground.prepare import (
-    PreparerConfig,
+from ml_playground.configuration import PreparerConfig
+from ml_playground.data_pipeline import (
+    create_standardized_metadata,
+    diff_file_states,
+    snapshot_file_states,
     split_train_val,
     write_bin_and_meta,
-    create_standardized_metadata,
-    snapshot_file_states,
-    diff_file_states,
 )
 from ml_playground.tokenizer import create_tokenizer
 from ml_playground.experiments.protocol import (
